series tests should be at test_reductions file
