+++ pandas/tests/reductions/test_reductions.py
@@ -1110,6 +1110,13 @@ class TestSeriesReductions:
             assert s.idxmax() == 0
             np.isnan(s.idxmax(skipna=False))
 
+    def test_sum_uint64(self):
+        # GH 53401
+        s = Series([10000000000000000000], dtype="uint64")
+        result = s.sum()
+        expected = np.uint64(10000000000000000000)
+        tm.assert_almost_equal(result, expected)
+
 
 class TestDatetime64SeriesReductions:
     # Note: the name TestDatetime64SeriesReductions indicates these tests
